Base, Core, Editor, Flags: fixed warnings

+++ modules/basegl/processors/volumeslicegl.cpp
@@ -581,7 +581,7 @@ void VolumeSliceGL::updateMaxSliceNumber() {
 
 void VolumeSliceGL::eventShiftSlice(Event* event) {
     auto wheelEvent = static_cast<WheelEvent*>(event);
-    int steps = wheelEvent->delta().y;
+    int steps = static_cast<int>(wheelEvent->delta().y);
     shiftSlice(steps);
     event->markAsUsed();
 }
